pyproject.toml
[project]
name = "polyjuice"
version = "0.1.0"
description = "Python scripts for Polymarket data collection and analysis"
authors = [
    {name = "Jonathan Tran"}
]
license = {text = "Apache 2.0"}
requires-python = ">=3.11"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "websockets (>=16.0,<17.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "sortedcontainers (>=2.4.0,<3.0.0)",
    "duckdb (>=1.5.0,<2.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "polars (>=1.39.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
